Reference Generation, Initial Estimate Generation Added
