+++ scripts/learning2slide/slide_observe_and_estimate/hw_kinova_sequential_control.py
