+++ scripts/learning2slide/slide_observe_and_estimate/hw_kinova_sequential_control_and_observe.py
@@ -29,7 +29,7 @@ from kinova_drake.kinova_station import (KinovaStationHardwareInterface, EndEffe
 from kinova_drake.controllers import (PSCommandSequenceController, PSCommandSequence, PartialStateCommand)
 from kinova_drake.observers import CameraViewer
 
-from object_tracker_system import ObjectTrackerSystem
+from object_tracker_system_friction_estimation import ObjectTrackerSystem
 
 import sequence_speed_limit_test
 import sequence_sliding_object
@@ -93,10 +93,10 @@ with KinovaStationHardwareInterface(n_dof) as station:
 
     ''' Command Sequence & Control '''
     # pscs, controller = sequence_speed_limit_test.command_sequence()
-    # pscs, controller = sequence_sliding_object.command_sequence()
-    # pscs, controller = sequence_sliding_back_and_forth.command_sequence()
     # pscs, controller = sequence_holding_object.command_sequence()
     # pscs, controller = sequence_pause.command_sequence()
+    # pscs, controller = sequence_sliding_object.command_sequence()
+    # pscs, controller = sequence_sliding_back_and_forth.command_sequence()
     pscs, controller = sequence_iterations_and_slide.command_sequence()
     
     controller = builder.AddSystem(controller)
@@ -175,8 +175,8 @@ with KinovaStationHardwareInterface(n_dof) as station:
         #     np.save(f, (ee_wrench_log.sample_times(), ee_wrench_log.data()[4,:]))
 
     if show_state_plots:
-        xmin = 10
-        xmax = 70
+        xmin = 185
+        xmax = 195
         
         six_dof_plots(ee_pose_log, "EE Pose", xmin, xmax,
                       save_state_plots, 'slide_data/ee_pose_data_plot.png')
