+++ scripts/learning2slide/slide_observe_and_estimate/sequence_pause.py
