+++ scripts/learning2slide/slide_observe_and_estimate/sequence_sliding_back_and_forth.py
